+++ source/helpers/locale.js
@@ -1,21 +1,21 @@
 /**
- * Whether can use `Intl.DateTimeFormat` for these `locales`.
- * Returns the first suitable one.
+ * From the list of `locales` it picks the first one
+ * that is supported by `Intl.DateTimeFormat`.
  * @param  {(string|string[])} locales
- * @return {?string} The first locale that can be used.
+ * @return {?string} The first locale that is supported, if any.
  */
-export function intlDateTimeFormatSupportedLocale(locales) {
+export function chooseOneSupportedLocale(locales) {
 	/* istanbul ignore else */
-	if (intlDateTimeFormatSupported()) {
+	if (isIntlDateTimeFormatSupported()) {
 		return Intl.DateTimeFormat.supportedLocalesOf(locales)[0]
 	}
 }
 
 /**
- * Whether can use `Intl.DateTimeFormat`.
+ * Whether the application can use `Intl.DateTimeFormat` class.
  * @return {boolean}
  */
-export function intlDateTimeFormatSupported() {
+export function isIntlDateTimeFormatSupported() {
 	// Babel transforms `typeof` into some "branches"
 	// so istanbul will show this as "branch not covered".
 	/* istanbul ignore next */