+++ source/helpers/locale.test.js
@@ -1,9 +1,9 @@
-import { intlDateTimeFormatSupportedLocale } from './locale.js'
+import { chooseOneSupportedLocale } from './locale.js'
 
 describe('locale', () => {
 	it(`should tell if can use Intl for date formatting`, () => {
-		expect(intlDateTimeFormatSupportedLocale('en')).to.equal('en')
-		expect(intlDateTimeFormatSupportedLocale('en-XX')).to.equal('en-XX')
-		expect(intlDateTimeFormatSupportedLocale(['en', 'ru'])).to.equal('en')
+		expect(chooseOneSupportedLocale('en')).to.equal('en')
+		expect(chooseOneSupportedLocale('en-XX')).to.equal('en-XX')
+		expect(chooseOneSupportedLocale(['en', 'ru'])).to.equal('en')
 	})
 })